Uptake core beta.ratebeer.com graphql API changes (#54)
* Update test to use new ratebeer graphql API (in progress)

Not all tests are updated, plus there is not support for searching for brewering as of this commit.

* Fix test_beer_unicode

The URL has changed with the updates to RateBeer.

* Fix test_beer_get_reviews

Ratings are now no longer required to have sub-ratings. Add a check to see if subratings exist before trying to parse.

* Remove commented out code from model refactor

* Switch to new API for #51

Still in the process of cleaning up tests. Reviews have not yet been looked at, but the API call has been extracted.

* Fix update after running tests #51

A few parts of the API have changed. In particular, weighted average is no longer available, and all 'blanks' are now set to None. Ex. 0 ratings or 0 ABV is now None, rather than a 0.

Additionally, a new exception JSONParseException has been added to make sure that when something strange comes down the pipes, we have something to catch it.